+++ crates/ruma-client-api/src/state.rs
@@ -1,5 +1,5 @@
 //! Endpoints for managing room state
 
+pub mod get_state_event_for_key;
 pub mod get_state_events;
-pub mod get_state_events_for_key;
 pub mod send_state_event;