+++ crates/ruma-common/CHANGELOG.md
@@ -57,6 +57,8 @@ Breaking changes:
   `key`, an enum that accepts any string.
   - The `key` field of `PushCondition::SenderNotificationPermission` uses the
     same type.
+- `RoomId::new()` was renamed to `RoomId::new_v1()`, as several formats are now
+  supported for this type.
 
 Bug fix:
 
@@ -102,6 +104,19 @@ Improvements:
   `m.room.redaction` event redacts.
 - Add `SpaceChildOrder` which allows to validate the `order` of an
   `m.space.child` event.
+- Add support for room version 12 and its unstable version `org.matrix.hydra.11`.
+- Add `explicitly_privilege_room_creators` and `additional_room_creators` to `AuthorizationRules`
+  to indicate whether room creators are considered to have "infinite" power level and whether
+  additional room creators can be specified with the `content.additional_creators` field of an
+  `m.room.create` event respectively.
+- Add `RoomPowerLevelsRules`, to provide tweaks to how the power level of a user is determined,
+  and add it to `PushConditionPowerLevelsCtx` via the `rules` field.
+- Add `room_id_format` to `RoomVersionRules`, to identify the format to use for room IDs depending
+  on the room version.
+- Add `RoomId::new_v2()` for the new format of room IDs. `RoomId::strip_sigil()` allows to access
+  the reference hash used in that format.
+- Add unstable support for in-app-only notifications as per MSC3768 under a new `unstable-msc3768`
+  feature.
 
 # 0.15.4
 
@@ -111,7 +126,6 @@ Bug fix:
   `serde_json::from_value` but not other functions like
   `serde_json::from_(str/slice)`. It now works with all 3 methods but is limited
   to deserializing JSON.
-
 
 # 0.15.3